Move ACI CNI validation to manifest code
- Set interface MTU to user provided MTU - 100. Default Neutron MTU is set to 1600, and thus interface MTU sets to 1500
- No default value for installerHostSubnet, error out if the user doesn't provide
