+++ pkg/asset/ignition/machine/master.go
@@ -44,7 +44,10 @@ func (a *Master) Generate(dependencies asset.Parents) error {
 
 	logrus.Info("Editing Master.........")
 
-	ignitionFiles := ign.IgnitionFiles(installConfig)
+	ignitionFiles, err := ign.IgnitionFiles(installConfig)
+	if err != nil {
+		return err
+	}
         for _, ignFile := range ignitionFiles {
                	a.Config.Storage.Files = append(a.Config.Storage.Files, ignFile)
        	}
