+++ pkg/asset/ignition/machine/master_test.go
@@ -91,7 +91,7 @@ func TestMasterGenerateCiscoAci(t *testing.T) {
 						KubeApiVLAN: "1021",
 						NeutronCIDR: ipnet.MustParseCIDR("5.6.7.8/5"),
 						InstallerHostSubnet: "9.10.11.12/10",
-						Mtu: "1600",
+						Mtu: "1500",
 					},
 				},
 			},
