+++ pkg/asset/ignition/machine/worker.go
@@ -44,7 +44,10 @@ func (a *Worker) Generate(dependencies asset.Parents) error {
 
 	logrus.Info("Editing Worker.........")
 
-        ignitionFiles := ign.IgnitionFiles(installConfig)
+        ignitionFiles, err := ign.IgnitionFiles(installConfig)
+	if err != nil {
+		return err
+	}
         for _, ignFile := range ignitionFiles {
                 a.Config.Storage.Files = append(a.Config.Storage.Files, ignFile)
         }
